requirements.txt
aniso8601==4.1.0
Click==7.0
coverage==4.5.2
Django==1.9.9
djangorestframework==3.4.1
Flask==1.0.2
Flask-API==1.1
Flask-RESTful==0.3.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
peewee==3.8.2
pytz==2018.9
six==1.12.0
Werkzeug==0.14.1
